+++ ui/src/main/java/com/layer/xdk/ui/DefaultXdkUiModule.java
@@ -6,6 +6,7 @@ import android.support.annotation.NonNull;
 import com.layer.sdk.LayerClient;
 import com.layer.xdk.ui.conversation.ConversationItemFormatter;
 import com.layer.xdk.ui.identity.IdentityFormatter;
+import com.layer.xdk.ui.message.model.MessageModelManager;
 import com.layer.xdk.ui.util.DateFormatter;
 import com.layer.xdk.ui.util.Log;
 import com.layer.xdk.ui.message.image.cache.ImageCacheWrapper;
@@ -97,5 +98,46 @@ public class DefaultXdkUiModule implements XdkUiModule {
     @Override
     public ConversationItemFormatter provideConversationItemFormatter() {
         return mServiceLocator.getConversationItemFormatter();
+    }
+
+    /**
+     * Provides the {@link MessageModelManager} from the {@link ServiceLocator} or creates one with
+     * the appropriate dependencies if not defined. This is not defined in the {@link XdkUiModule}
+     * interface as instances can be created automatically without a provider due to the annotated
+     * constructor. Should a custom instance be created it would be overridden in the custom module.
+     *
+     * @return the {@link MessageModelManager} of the service locator or a new instance using
+     * dependencies from the service locator
+     */
+    @Provides
+    @Singleton
+    @NonNull
+    public MessageModelManager provideMessageModelManager() {
+        if (mServiceLocator.getMessageModelManager() != null) {
+            return mServiceLocator.getMessageModelManager();
+        }
+
+        if (mServiceLocator.getAppContext() == null) {
+            if (Log.isLoggable(Log.ERROR)) {
+                Log.e("No Context when attempting to create a MessageModelManager");
+            }
+            throw new IllegalStateException(
+                    "No Context when attempting to create a MessageModelManager");
+        }
+        if (mServiceLocator.getLayerClient() == null) {
+            if (Log.isLoggable(Log.ERROR)) {
+                Log.e("No LayerClient when attempting to create a MessageModelManager");
+            }
+            throw new IllegalStateException(
+                    "No LayerClient when attempting to create a MessageModelManager");
+        }
+
+        // Create a new object using dependencies from the service locator since there is no
+        // way to automatically create the object via Dagger here.
+        return new MessageModelManager(mServiceLocator.getAppContext(),
+                mServiceLocator.getLayerClient(),
+                mServiceLocator.getIdentityFormatter(),
+                mServiceLocator.getDateFormatter(),
+                mServiceLocator.getImageCacheWrapper());
     }
 }